Enable version specific sourceSets for compiler-test module
